hitw blocker

+++ TheRegularEngineV2/TheRegularEditor/src/EditorAssetManager.cpp
@@ -96,8 +96,6 @@ namespace TRE
 
 	void AssetManager::RemoveAsset(const std::string& assetName)
 	{
-		if (m_AssetNameToHandle[assetName].second == false)
-			return;
 		const ResourceHandle handle = m_AssetNameToHandle[assetName].first;
 		ResourceType assetType = ResourceManager::Instance().GetResourceType(handle);
 		std::filesystem::path assetsPath = "../Assets/";
@@ -111,8 +109,11 @@ namespace TRE
 			assetsPath = "../Assets/";
 			assetsPath += assetName;
 			std::filesystem::remove(assetsPath);
-			resourcePath += hexHandle + ".geom";
-			std::filesystem::remove(resourcePath);
+			if (m_AssetNameToHandle[assetName].second)
+			{
+				resourcePath += hexHandle + ".geom";
+				std::filesystem::remove(resourcePath);
+			}
 		}
 		else if (assetType == ResourceType::Texture)
 		{
@@ -121,15 +122,21 @@ namespace TRE
 			assetsPath = "../Assets/";
 			assetsPath += assetName;
 			std::filesystem::remove(assetsPath);
-			resourcePath += hexHandle + ".DDS";
-			std::filesystem::remove(resourcePath);
+			if (m_AssetNameToHandle[assetName].second)
+			{
+				resourcePath += hexHandle + ".DDS";
+				std::filesystem::remove(resourcePath);
+			}
 		}
 		else if (assetType == ResourceType::Material)
 		{
-			assetsPath += hexHandle + ".material" + ".desc";
-			std::filesystem::remove(assetsPath);
-			resourcePath += hexHandle + ".material";
-			std::filesystem::remove(resourcePath);
+			if (m_AssetNameToHandle[assetName].second)
+			{
+				assetsPath += hexHandle + ".material" + ".desc";
+				std::filesystem::remove(assetsPath);
+				resourcePath += hexHandle + ".material";
+				std::filesystem::remove(resourcePath);
+			}
 		}
 		ResourceManager::Instance().RemoveResource(m_AssetNameToHandle[assetName].first);
 		m_AssetNameToHandle.erase(assetName);
